Update JamfObjectUploaderBase.py - handle _register object types
Processor failed when used against `jamf_protect_register` as was attempting to get an object ID with get_api_obj_id_from_name. This object type does not have an ID or name and always expects a payload to be POSTed to it.

Should be handled correctly with this change (have tested with `_settings` `_command` and `os_x_configuration_profile` object types and all upload as expected).