Move the `curl` `headers_file` to each session's temp dir
The `headers_file` must remain in the unique directory for each session as that file contains the return content for each individual `curl` call, specifically the HTTP status code, which is checked before moving forward.

Using the same file for all sessions means the status code is being overwritten by different sessions' `curl` calls and when the `headers_file` is read back in, it may get the HTTP status code of a different sessions' `curl` execution.

This fixes/reverts a change in #247.